Add beginnings of AxisController class
As labelled, create an AxisController class and Arduino library, so that
this project harness the full power of OOP, and different servos, each
controlling a seperate axis on the robotic arm, can be instantiated as
needed.

Note that, for now, this requires the SoftwareSerial library, and therefore
a piece of hardware with serial communications capability.

Signed-off-by: Boggs <boggsj@gmail.com>