Finish proto-type AxisController
I kinda-sorta finished the AxisContoller library. That is, I finished
everything I've though of, but it isn't tested yet. All it's supposed to do
is read in input from the serial in, pare it out into 10 different pieces
of data, and convert one of those pieces of data to a servo position. The
input should be sensor data from a Myo armband.

Signed-off-by: Boggs <boggsj@gmail.com>